# docs/swagger.yaml
definitions:
  dopTypes.ErrRep:
    properties:
      desc:
        type: string
      error_code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
  entities.JwkSetSt:
    properties:
      keys:
        items:
          $ref: '#/definitions/entities.JwkSt'
        type: array
    type: object
  entities.JwkSt:
    properties:
      alg:
        type: string
      e:
        type: string
      kid:
        type: string
      kty:
        type: string
      "n":
        type: string
      use:
        type: string
    type: object
  entities.JwtCreateRepSt:
    properties:
      token:
        type: string
    type: object
  entities.JwtCreateReqSt:
    properties:
      exp_seconds:
        type: integer
      payload:
        additionalProperties:
          type: any
        type: object
      sub:
        type: string
    type: object
  entities.JwtValidateRepSt:
    properties:
      claims:
        additionalProperties:
          type: any
        type: object
      valid:
        type: boolean
    type: object
  entities.JwtValidateReqSt:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /jwk/set:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.JwkSetSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - jwk
  /jwt:
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.JwtCreateReqSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.JwtCreateRepSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - jwt
  /jwt/validate:
    put:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.JwtValidateReqSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.JwtValidateRepSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - jwt
swagger: "2.0"